requirements.txt
google-generativeai
pydub
tqdm 